pyproject.toml
[project]
name = "KWT"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django",
    "psycopg[binary,pool]",
    "bg-django-mobile-detector",
    "django-ratelimit",
    "django-storages[boto3]",
    "bg-auth",
    "bg-uuidv7",
    "django-phonenumber-field[phonenumbers]",
    "python-dotenv",
    "django_crontab",
    "valkey",
    "django-valkey",
    "celery",
    "djangorestframework",
    "django-cors-headers",
    "drf-spectacular",
    "pywebpush",
    "spotdl",
]

pytest_plugins = ["configs.pytest_plugins.django_projects"]

[dependency-groups]
celery = ["django-celery-beat"]
uwsgi = [
    "Pillow",
    "pyuwsgi",
    "bg-django-caching-paginator",
    "django-cryptography",
    "django-prometheus",
]
daphne = ["daphne", "channels", "channels-redis"]
dev = [
    "Pillow",
    "pyuwsgi; sys_platform == 'linux'",
    "bg-django-caching-paginator",
    "django-cryptography",
    "django-prometheus",
    "daphne",
    "channels",
    "channels-redis",
]

[tool.uv.sources]
bg-auth = { workspace = true }
bg-django-caching-paginator = { workspace = true }
bg-django-mobile-detector = { workspace = true }
bg-uuidv7 = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["korfbal", "apps"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "korfbal.settings_test"
python_files = ["test_*.py"]
pythonpath = ["/Users/butrosgroot/Documents/MonoRepo"]
pytest_plugins = ["configs.pytest_plugins.django_projects"]
